# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/work-schedule": {
      "post": {
        "operationId": "WorkScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agenda criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSchedule"
                }
              }
            }
          }
        },
        "summary": "Criar uma nova agenda de trabalho",
        "tags": [
          "Work Schedule"
        ]
      },
      "get": {
        "operationId": "WorkScheduleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de agendas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkSchedule"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todas as agendas de trabalho",
        "tags": [
          "Work Schedule"
        ]
      }
    },
    "/work-schedule/{id}": {
      "get": {
        "operationId": "WorkScheduleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agenda encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSchedule"
                }
              }
            }
          }
        },
        "summary": "Buscar agenda por ID",
        "tags": [
          "Work Schedule"
        ]
      },
      "patch": {
        "operationId": "WorkScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agenda atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSchedule"
                }
              }
            }
          }
        },
        "summary": "Atualizar agenda",
        "tags": [
          "Work Schedule"
        ]
      },
      "delete": {
        "operationId": "WorkScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agenda removida"
          }
        },
        "summary": "Remover agenda",
        "tags": [
          "Work Schedule"
        ]
      }
    },
    "/work-schedule-day": {
      "post": {
        "operationId": "WorkScheduleDayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkScheduleDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dia da semana criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkScheduleDay"
                }
              }
            }
          }
        },
        "summary": "Criar um novo dia da agenda",
        "tags": [
          "Work Schedule Day"
        ]
      },
      "get": {
        "operationId": "WorkScheduleDayController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de dias da semana",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkScheduleDay"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os dias da agenda",
        "tags": [
          "Work Schedule Day"
        ]
      }
    },
    "/work-schedule-day/work-schedule/{workScheduleId}": {
      "get": {
        "operationId": "WorkScheduleDayController_findByWorkScheduleId",
        "parameters": [
          {
            "name": "workScheduleId",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dias da semana da agenda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkScheduleDay"
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar dias por agenda",
        "tags": [
          "Work Schedule Day"
        ]
      }
    },
    "/work-schedule-day/{id}": {
      "get": {
        "operationId": "WorkScheduleDayController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do dia",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dia encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkScheduleDay"
                }
              }
            }
          }
        },
        "summary": "Buscar dia por ID",
        "tags": [
          "Work Schedule Day"
        ]
      },
      "patch": {
        "operationId": "WorkScheduleDayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do dia",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkScheduleDayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dia da semana atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkScheduleDay"
                }
              }
            }
          }
        },
        "summary": "Atualizar dia",
        "tags": [
          "Work Schedule Day"
        ]
      },
      "delete": {
        "operationId": "WorkScheduleDayController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do dia",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dia removido"
          }
        },
        "summary": "Remover dia",
        "tags": [
          "Work Schedule Day"
        ]
      }
    },
...
    "/forecast/work-schedule/{workScheduleId}": {
      "get": {
        "operationId": "ForecastController_findByWorkSchedule",
        "parameters": [
          {
            "name": "workScheduleId",
            "required": true,
            "in": "path",
            "description": "ID da agenda de trabalho",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Previsões da agenda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Forecast"
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar previsões por agenda de trabalho",
        "tags": [
          "Forecast"
        ]
      }
    },
...
  },
  "info": {
    "title": "Work Schedule API",
    "description": "API para gerenciamento de agendas de trabalho",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateWorkScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "WorkSchedule": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWorkScheduleDayDto": {
        "type": "object",
        "properties": {}
      },
      "WorkScheduleDay": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkScheduleDayDto": {
        "type": "object",
        "properties": {}
      },
...
      "Forecast": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}